changes to spring pages, measurement adding

+++ app/Http/Controllers/SpringController.php
@@ -2,9 +2,15 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Photo;
 use App\Models\Spring;
 use App\Models\SpringDatabaseLink;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Validation\Rule;
+use Inertia\Inertia;
 use function Psy\debug;
 use Illuminate\Support\Facades\Auth;
 use App\Models\SpringReference;
@@ -18,19 +24,55 @@ class SpringController extends Controller
      */
     public function index()
     {
-        $springs = Spring::all();
-        return view('springs.index',compact('springs'));
-            //->with('i', (request()->input('page', 1) - 1) * 5);
+        $springs = Spring::where('status', ['submitted', 'confirmed'])->get();
+        return Inertia::render('Springs/Index', ['springs' => $springs]);
     }
 
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Inertia\Response
      */
     public function create()
     {
-        return view('springs.create');
+        if (Auth::user()) {
+            return Inertia::render('Springs/Create', [
+                'classifications' => SpringController::getClassifications(),
+                'ownerships' => SpringController::getOwnerships(),
+                'statuses' => SpringController::getStatuses()
+            ]);
+        }
+        $springs = Spring::where('status', ['submitted', 'confirmed'])->get();
+        return Inertia::render('Springs/Index', ['springs' => $springs]);
+    }
+
+    public function getClassifications() {
+        return [
+            array( 'id' => 'rheocrene', 'name' => 'Rheocrene'),
+            array('id' => 'hillslope_spring', 'name' => 'Hillslope spring'),
+            array('id' => 'limnocrene', 'name' => 'Limnocrene'),
+            array('id' => 'helocrene', 'name' => 'Helocrene'),
+            array('id' => 'cave_spring', 'name' => 'Cave spring'),
+            array('id' => 'hypocrene', 'name' => 'Hypocrene'),
+            array('id' => 'captured_spring', 'name' => 'Captured spring'),
+            array('id' => 'karst_spring', 'name' => 'Karst spring'),
+        ];
+    }
+
+    public function getOwnerships() {
+        return [
+            array('id' => 'private_property', 'name' => 'Private property'),
+            array('id' => 'state_property', 'name' => 'State property'),
+            array('id' => 'municipal_property', 'name' => 'Municipal property'),
+        ];
+    }
+
+    public function getStatuses() {
+        return [
+            array('id' => 'draft', 'name' => 'Draft'),
+            array('id' => 'submitted', 'name' => 'Submitted'),
+            array('id' => 'confirmed', 'name' => 'Confirmed'),
+        ];
     }
 
     /**
@@ -41,48 +83,73 @@ class SpringController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
+        Validator::make($request->all(), [
+            'description' => ['required', 'string', 'max:255'],
             'latitude' => 'required',
             'longitude' => 'required',
-            'description' => 'required'
-        ]);
+        ])->validateWithBag('addSpring');
 
         $request['user_id'] = Auth::id();
-        $spring = Spring::create($request->all());
-        // save references
-        //$reference = new Reference();
-        //Reference::create($request);
-        //print_r($request['spring_references']);
-        //var_dump($request['spring_references']);exit;
-        foreach ($request['spring_references'] as $reference_info) {
-            //var_dump($reference_info);exit;
-            $spring_reference = new SpringReference();
-            if ($reference_info['url']) {
-                $spring_reference->spring_id = $spring->id;
-                $spring_reference->url = $reference_info['url'];
-                if ($reference_info['url_title']) {
-                    $spring_reference->url_title = $reference_info['url_title'];
-                }
-                $spring_reference->save();
-            }
-        }
-        // save database links
-        foreach ($request['spring_databases'] as $database_info) {
-            if (empty($database_info['database_name']) && empty($database_info['code'])
-                && empty($database_info['spring_name']) && empty($database_info['url'])) {
-                continue;
-            }
-            $spring_database_link = new SpringDatabaseLink();
-            $spring_database_link->spring_id = $spring->id;
-            $spring_database_link->database_name = $database_info['database_name'];
-            $spring_database_link->code = $database_info['code'];
-            $spring_database_link->spring_name = $database_info['spring_name'];
-            $spring_database_link->url = $database_info['url'];
-            $spring_database_link->save();
-        }
+        // create wateract code
+        $new_id = DB::table('springs')->max('id') + 1;
+        var_dump(sprintf("%05s", $new_id));
+        $format = 'There are %d monkeys in the %s';
+        echo sprintf('%s%05d', $request['country'], $new_id);
+        $code = sprintf('%s%05d', $request['country'], $new_id);
 
+        $request['code'] = $code;
+        $spring = Spring::create($request->all());
+
+        SpringController::saveReferences($spring, $request['references']);
+        SpringController::saveDatabaseLinks($spring, $request['database_links']);
+
+        if (!empty($request['photos'])) {
+            foreach ($request['photos'] as $photo_raw) {
+                var_dump($photo_raw);
+                $photo = new Photo();
+                $photo->spring_id = $spring->id;
+                $path = Storage::disk('public')->put('spring-photos', $photo_raw);
+                $photo->path = $path;
+                $photo->save();
+            }
+        }
         return redirect()->route('springs.index')
             ->with('success','Spring created successfully.');
+    }
+
+    public function saveReferences($spring, $references_info) {
+        if (!empty($references_info)) {
+            foreach ($references_info as $reference_info) {
+                $spring_reference = new SpringReference();
+                if ($reference_info['url']) {
+                    $spring_reference->spring_id = $spring['id'];
+                    $spring_reference->url = $reference_info['url'];
+                    if ($reference_info['url_title']) {
+                        $spring_reference->url_title = $reference_info['url_title'];
+                    }
+                    $spring_reference->save();
+                }
+            }
+        }
+    }
+
+    public function saveDatabaseLinks($spring, $databases_info) {
+        if (isset($databases_info)) {
+            //$databases = json_decode($databases_info);
+            foreach ($databases_info as $database_info) {
+                if (empty($database_info['database_name']) && empty($database_info['code'])
+                    && empty($database_info['spring_name']) && empty($database_info['url'])) {
+                    continue;
+                }
+                $spring_database_link = new SpringDatabaseLink();
+                $spring_database_link->spring_id = $spring->id;
+                $spring_database_link->database_name = $database_info['database_name'];
+                $spring_database_link->code = $database_info['code'];
+                $spring_database_link->spring_name = $database_info['spring_name'];
+                $spring_database_link->url = $database_info['url'];
+                $spring_database_link->save();
+            }
+        }
     }
 
     /**
@@ -93,7 +160,8 @@ class SpringController extends Controller
      */
     public function show(Spring $spring)
     {
-        return view('springs.show', compact('spring'));
+        $spring = Spring::where('id', $spring->id)->with('references')->with('database_links')->first();
+        return Inertia::render('Springs/Show', ['spring' => $spring]);
     }
 
     /**
@@ -104,7 +172,14 @@ class SpringController extends Controller
      */
     public function edit(Spring $spring)
     {
-        return view('springs.edit', compact('spring'));
+        $spring = Spring::where('id', $spring->id)->with('references')->with('database_links')->first();
+        if (Auth::user()) {
+            return Inertia::render('Springs/Edit', [
+                'classifications' => SpringController::getClassifications(),
+                'ownerships' => SpringController::getOwnerships(),
+                'spring' => $spring]);
+        }
+        return Inertia::render('Springs/Show', ['spring' => $spring]);
     }
 
     /**
@@ -116,6 +191,8 @@ class SpringController extends Controller
      */
     public function update(Request $request, Spring $spring)
     {
+        $this->authorize('update', $spring);
+
         $request->validate([
             'latitude' => 'required',
             'longitude' => 'required',
@@ -123,7 +200,7 @@ class SpringController extends Controller
         ]);
         $spring->update($request->all());
 
-        foreach ($request['spring_references'] as $reference_info) {
+        foreach ($request['references'] as $reference_info) {
             if (isset($reference_info['id'])) {
                 //$spring_reference = SpringReference::where('id' , '=' , $reference_info['id'] )->get();
                 $spring_reference = SpringReference::find($reference_info['id']);
@@ -149,6 +226,34 @@ class SpringController extends Controller
 
         }
 
+        foreach ($request['spring_databases'] as $database_info) {
+            if (isset($database_info['id'])) {
+                $spring_database_link = SpringDatabaseLink::find($database_info['id']);
+                if (empty($database_info['database_name']) && empty($database_info['code'])
+                    && empty($database_info['spring_name']) && empty($database_info['url'])) {
+                    $spring_database_link->delete();
+                }
+                $spring_database_link->database_name = $database_info['database_name'];
+                $spring_database_link->code = $database_info['code'];
+                $spring_database_link->spring_name = $database_info['spring_name'];
+                $spring_database_link->url = $database_info['url'];
+                $spring_database_link->save();
+            } else {
+                if (empty($database_info['database_name']) && empty($database_info['code'])
+                    && empty($database_info['spring_name']) && empty($database_info['url'])) {
+                    continue;
+                }
+                $spring_database_link = new SpringDatabaseLink();
+                $spring_database_link->spring_id = $spring->id;
+                $spring_database_link->database_name = $database_info['database_name'];
+                $spring_database_link->code = $database_info['code'];
+                $spring_database_link->spring_name = $database_info['spring_name'];
+                $spring_database_link->url = $database_info['url'];
+                $spring_database_link->save();
+            }
+
+        }
+
         return redirect()->route('springs.show', compact('spring'))
             ->with('success','Spring updated successfully.');
     }
@@ -162,6 +267,8 @@ class SpringController extends Controller
      */
     public function destroy(Spring $spring)
     {
+        $this->authorize('delete', $spring);
+
         $spring->delete();
 
         return redirect()->route('springs.index')
